move demo, anmotation remote.rs and obeserver.java
