+++ java/connector-node/risingwave-connector-service/src/main/java/com/risingwave/connector/SinkStreamObserver.java
@@ -138,21 +138,27 @@ public class SinkStreamObserver implements StreamObserver<ConnectorServiceProto.
                                     .asRuntimeException();
                         }
                         break;
-                    case STREAM_CHUNK_PAYLOAD:
-                        if (deserializer == null) {
-                            deserializer = new StreamChunkDeserializer(tableSchema);
-                        }
-
-                        if (deserializer instanceof StreamChunkDeserializer) {
-                            rows =
-                                    deserializer.deserialize(
-                                            sinkTask.getWrite().getStreamChunkPayload());
-                        } else {
-                            throw INTERNAL.withDescription(
-                                            "invalid payload type: expected StreamChunk, got "
-                                                    + deserializer.getClass().getName())
-                                    .asRuntimeException();
-                        }
+                        //                    case STREAM_CHUNK_PAYLOAD:
+                        //                        if (deserializer == null) {
+                        //                            deserializer = new
+                        // StreamChunkDeserializer(tableSchema);
+                        //                        }
+                        //
+                        //                        if (deserializer instanceof
+                        // StreamChunkDeserializer) {
+                        //                            rows =
+                        //                                    deserializer.deserialize(
+                        //
+                        // sinkTask.getWrite().getStreamChunkPayload());
+                        //                        } else {
+                        //                            throw INTERNAL.withDescription(
+                        //                                            "invalid payload type:
+                        // expected StreamChunk, got "
+                        //                                                    +
+                        // deserializer.getClass().getName())
+                        //                                    .asRuntimeException();
+                        //                        }
+                        //                        break;
                     default:
                         throw INVALID_ARGUMENT
                                 .withDescription("invalid payload type")
