+++ java/java-binding/src/main/java/com/risingwave/java/binding/StreamChunkDeserializer.java
@@ -12,38 +12,23 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.risingwave.connector;
+package com.risingwave.java.binding;
 
 import static io.grpc.Status.INVALID_ARGUMENT;
 
+import com.risingwave.connector.Deserializer;
 import com.risingwave.connector.api.TableSchema;
 import com.risingwave.connector.api.sink.ArraySinkrow;
 import com.risingwave.connector.api.sink.SinkRow;
-import com.risingwave.java.binding.StreamChunkIterator;
-import com.risingwave.java.binding.StreamChunkRow;
 import com.risingwave.proto.ConnectorServiceProto.SinkStreamRequest.WriteBatch.StreamChunkPayload;
 import com.risingwave.proto.Data;
-import java.util.ArrayList;
 import java.util.Iterator;
-import java.util.List;
 
 public class StreamChunkDeserializer implements Deserializer {
     private final TableSchema tableSchema;
 
     public StreamChunkDeserializer(TableSchema tableSchema) {
         this.tableSchema = tableSchema;
-    }
-
-    public Iterator<SinkRow> deserializeByIterator(Object payload) {
-        if (!(payload instanceof StreamChunkPayload)) {
-            throw INVALID_ARGUMENT
-                    .withDescription(
-                            "expected StreamChunkPayload, got " + payload.getClass().getName())
-                    .asRuntimeException();
-        }
-        StreamChunkPayload streamChunkPayload = (StreamChunkPayload) payload;
-        return new MyIterator(
-                tableSchema, new StreamChunkIterator(streamChunkPayload.toByteArray()));
     }
 
     @Override
@@ -55,29 +40,16 @@ public class StreamChunkDeserializer implements Deserializer {
                     .asRuntimeException();
         }
         StreamChunkPayload streamChunkPayload = (StreamChunkPayload) payload;
-        List<SinkRow> sinkRowList = new ArrayList<>();
-        try (StreamChunkIterator iter = new StreamChunkIterator(streamChunkPayload.toByteArray())) {
-            while (true) {
-                try (StreamChunkRow row = iter.next()) {
-                    if (row == null) {
-                        break;
-                    }
-                    Object[] values = new Object[tableSchema.getNumColumns()];
-                    for (String columnName : tableSchema.getColumnNames()) {
-                        int columnIdx = tableSchema.getColumnIndex(columnName);
-                        Data.DataType.TypeName typeName = tableSchema.getColumnType(columnName);
-                        values[tableSchema.getColumnIndex(columnName)] =
-                                validateStreamChunkDataTypes(typeName, columnIdx, row);
-                    }
-                    sinkRowList.add(new ArraySinkrow(row.getOp(), values));
-                }
-            }
-        }
-        return sinkRowList.iterator();
+        return new MyIterator(
+                tableSchema,
+                new StreamChunkIterator(streamChunkPayload.getBinaryData().toByteArray()));
     }
 
     private static Object validateStreamChunkDataTypes(
             Data.DataType.TypeName typeName, int columnIdx, StreamChunkRow row) {
+        if (row.isNull(columnIdx)) {
+            return null;
+        }
         switch (typeName) {
             case INT16:
                 return row.getShort(columnIdx);
