+++ src/connector/src/sink/remote.rs
@@ -17,19 +17,19 @@ use std::time::Duration;
 
 use anyhow::anyhow;
 use async_trait::async_trait;
-use prost::Message;
 use risingwave_common::array::StreamChunk;
 #[cfg(test)]
 use risingwave_common::catalog::Field;
 use risingwave_common::catalog::Schema;
 use risingwave_common::config::{MAX_CONNECTION_WINDOW_SIZE, STREAM_WINDOW_SIZE};
+use risingwave_common::row::Row;
 #[cfg(test)]
 use risingwave_common::types::DataType;
 use risingwave_common::types::{DatumRef, ScalarRefImpl};
+use risingwave_common::util::iter_util::ZipEqFast;
 use risingwave_pb::connector_service::connector_service_client::ConnectorServiceClient;
-use risingwave_pb::connector_service::sink_stream_request::write_batch::{
-    Payload, StreamChunkPayload,
-};
+use risingwave_pb::connector_service::sink_stream_request::write_batch::json_payload::RowOp;
+use risingwave_pb::connector_service::sink_stream_request::write_batch::{JsonPayload, Payload};
 use risingwave_pb::connector_service::sink_stream_request::{
     Request as SinkRequest, StartEpoch, StartSink, SyncBatch, WriteBatch,
 };
@@ -236,8 +236,25 @@ impl<const APPEND_ONLY: bool> RemoteSink<APPEND_ONLY> {
 #[async_trait]
 impl<const APPEND_ONLY: bool> Sink for RemoteSink<APPEND_ONLY> {
     async fn write_batch(&mut self, chunk: StreamChunk) -> Result<()> {
-        let prost_stream_chunk = chunk.to_protobuf();
-        let binary_data = Message::encode_to_vec(&prost_stream_chunk);
+        let mut row_ops = vec![];
+        for (op, row_ref) in chunk.rows() {
+            let mut map = serde_json::Map::new();
+            row_ref
+                .iter()
+                .zip_eq_fast(self.schema.fields.iter())
+                .for_each(|(v, f)| {
+                    map.insert(f.name.clone(), parse_datum(v));
+                });
+            let row_op = RowOp {
+                op_type: op.to_protobuf() as i32,
+                line: serde_json::to_string(&map)
+                    .map_err(|e| SinkError::Remote(format!("{:?}", e)))?,
+            };
+
+            row_ops.push(row_op);
+        }
+        // let prost_stream_chunk = chunk.to_protobuf();
+        // let binary_data = Message::encode_to_vec(&prost_stream_chunk);
 
         let epoch = self.epoch.ok_or_else(|| {
             SinkError::Remote("epoch has not been initialize, call `begin_epoch`".to_string())
@@ -248,7 +265,8 @@ impl<const APPEND_ONLY: bool> Sink for RemoteSink<APPEND_ONLY> {
                 request: Some(SinkRequest::Write(WriteBatch {
                     epoch,
                     batch_id,
-                    payload: Some(Payload::StreamChunkPayload(StreamChunkPayload{binary_data})),
+                    payload: Some(Payload::JsonPayload(JsonPayload { row_ops })),
+                    // payload: Some(Payload::JsonPayload(JsonPayload { row_ops })),
                 })),
             })
             .map_err(|e| SinkError::Remote(e.to_string()))?;
