+++ src/java_binding/make-java-binding.toml
@@ -96,5 +96,8 @@ cd ${RISINGWAVE_ROOT}/java
     java -cp "./java-binding-integration-test/target/dependency/*:./java-binding-integration-test/target/classes" \
      -Djava.library.path=${RISINGWAVE_ROOT}/target/debug  com.risingwave.java.binding.StreamChunkDemo
 
+(cd ${RISINGWAVE_ROOT} && cargo run --bin data-chunk-payload-generator) | \
+    java -cp "./java-binding-integration-test/target/dependency/*:./java-binding-integration-test/target/classes" \
+     -Djava.library.path=${RISINGWAVE_ROOT}/target/debug  com.risingwave.java.binding.StreamChunkDeserializerDemo
 
 '''
